DataRetriveService/requirements.txt
alembic==1.14.1
# apache-airflow==2.10.5
fastapi==0.115.8
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==1.4.54
asyncpg==0.30.0
psycopg2-binary==2.9.10
uvicorn==0.34.0